Add files via upload
I was able to make the text clickable, but there is now a styling issue where the text is underlined red when not the cursor is not hovering over it. I tried to to find what part of the stylesheet is causing this underlining to happen, but wasn't able to. Hopefully we can go over this in the meeting :)